change version. deleted igonore-xx.txt #24
