+++ seiko.smartcs/meta/runtime.yml
@@ -1,5 +1,5 @@
 ---
-requires_ansible: '>=2.14.0,<2.17'
+requires_ansible: '>=2.15.0,<2.18'
 plugin_routing:
   action:
     smartcs_command:
