+++ seiko.smartcs/plugins/action/smartcs_tty_command.py
@@ -1,1 +1,1 @@
-smartcs.py
+smartcs_config.py
