--- seiko.smartcs/tests/sanity/ignore-2.10.txt
@@ -1,1 +1,0 @@
-plugins/action/smartcs.py action-plugin-docs # base class for deprecated network platform modules using `connection: local`
